atusb: fw: Provide TRAC status
Upon Tx done condition, returning the sequence number is useful but we
might also return the TRAC value which is needed to ensure that the
packet we sent got ACKed.

We then need to read the TRAC status register upon Tx completion and
send this information to the atusb Linux driver as part of the status
message. First byte remains the sequence number for ensuring backward
compatibility, a second byte is added to forward the TRAC register
status.

We need to move the transition to RX_AACK_ON after reading the trac
register value. The current optimization by switchting to RX_AACK_ON
after invoking transceiver transmission could have the side effect that
the TRAC register is not read out before a frame is received. Receiving
another frame will overwrite the TRAC register. We can only switch to
RX_AACK_ON state after we read out the TRAC register.

Signed-off-by: Alexander Aring <aahringo@redhat.com>
Tested-by: Miquel Raynal <miquel.raynal@bootlin.com>
[Miquel Raynal: Moved the data array out of the stack, rewrote commit log]
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>