atusb: fw: Makefile: add comment on how to fix build for avr-gcc > 12
Starting with avr-gcc 12 we need to tell the compiler clearly that we
are a firmwware and we start at address 0 to avoid warnings and errors
of the firmware build. Sadly this new option can not be enabled by
default. Older gcc versions do not have this commandline options and
would break (they would also succeed in compiling the firmware).
So we do the next big thing and at least add a comment with the fix in
the Makefile to get started.
Once the gcc version is more established we can change the defaults and
enable it.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>